+++ pnet-data-api-java/src/main/java/pnet/data/api/apache5/Apache5HttpClientBasedRestCallFactoryConfig.java
@@ -2,37 +2,35 @@ package pnet.data.api.apache5;
 
 import at.porscheinformatik.happyrest.RestCallFactory;
 import at.porscheinformatik.happyrest.RestLoggerAdapter;
-import at.porscheinformatik.happyrest.SystemRestLoggerAdapter;
 import at.porscheinformatik.happyrest.apache5.Apache5RestCallFactory;
-import at.porscheinformatik.happyrest.slf4j.Slf4jRestLoggerAdapter;
 import at.porscheinformatik.happyrest.spring.SpringRestFormatter;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.Optional;
-import org.springframework.beans.factory.annotation.Qualifier;
+import java.util.Set;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
 import org.springframework.core.convert.ConversionService;
+import org.springframework.core.convert.converter.Converter;
 import pnet.data.api.client.PnetDataRestCallFactoryConfig;
+import pnet.data.api.client.PnetDataRestCallFactoryConfigurer;
 import pnet.data.api.util.PnetDataApiUtils;
 
 @Configuration
-@Import({ PnetDataRestCallFactoryConfig.class })
-public class Apache5HttpClientBasedRestCallFactoryConfig {
+public class Apache5HttpClientBasedRestCallFactoryConfig extends PnetDataRestCallFactoryConfig {
+
+    protected Apache5HttpClientBasedRestCallFactoryConfig(
+        Optional<PnetDataRestCallFactoryConfigurer> configurer,
+        Optional<RestLoggerAdapter> loggerAdapter,
+        Optional<Set<? extends Converter<?, ?>>> attributeConverters
+    ) {
+        super(configurer, loggerAdapter, attributeConverters);
+    }
 
     @Bean
-    public RestCallFactory restCallFactory(
-        @Qualifier("pnetDataApiObjectMapper") ObjectMapper objectMapper,
-        @Qualifier("pnetDataApiConversionService") ConversionService conversionService,
-        @Qualifier("pnetDataApiRestLoggerAdapter") Optional<RestLoggerAdapter> optionalLoggerAdapter
-    ) {
-        RestLoggerAdapter loggerAdapter = optionalLoggerAdapter.orElseGet(() -> {
-            if (Slf4jRestLoggerAdapter.isSlf4jAvailable()) {
-                return Slf4jRestLoggerAdapter.getDefault();
-            }
-
-            return SystemRestLoggerAdapter.INSTANCE;
-        });
+    public RestCallFactory apache5HttpClientBasedRestCallFactory() {
+        ObjectMapper objectMapper = objectMapper();
+        RestLoggerAdapter loggerAdapter = restLoggerAdapter();
+        ConversionService conversionService = conversionService();
 
         return Apache5RestCallFactory.create(loggerAdapter, objectMapper)
             .withUserAgent(PnetDataApiUtils.getUserAgent("Apache's HttpClient"))