+++ pnet-data-api-java/src/main/java/pnet/data/api/client/PnetDataRestCallFactoryConfig.java
@@ -1,44 +1,92 @@
 package pnet.data.api.client;
 
+import at.porscheinformatik.happyrest.RestLoggerAdapter;
+import at.porscheinformatik.happyrest.SystemRestLoggerAdapter;
+import at.porscheinformatik.happyrest.slf4j.Slf4jRestLoggerAdapter;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
+import java.util.HashSet;
 import java.util.Locale;
 import java.util.Optional;
 import java.util.Set;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.context.support.ConversionServiceFactoryBean;
 import org.springframework.core.convert.ConversionService;
 import org.springframework.core.convert.converter.Converter;
 import pnet.data.api.GeoDistance;
 import pnet.data.api.client.jackson.JacksonPnetDataApiModule;
 
-@Configuration
-public class PnetDataRestCallFactoryConfig {
+public abstract class PnetDataRestCallFactoryConfig {
 
-    @Bean
-    public ObjectMapper pnetDataApiObjectMapper() {
+    private final Optional<PnetDataRestCallFactoryConfigurer> configurer;
+    private final Optional<RestLoggerAdapter> loggerAdapter;
+    private final Optional<Set<? extends Converter<?, ?>>> attributeConverters;
+
+    protected PnetDataRestCallFactoryConfig(
+        Optional<PnetDataRestCallFactoryConfigurer> configurer,
+        Optional<RestLoggerAdapter> loggerAdapter,
+        Optional<Set<? extends Converter<?, ?>>> attributeConverters
+    ) {
+        this.configurer = configurer;
+        this.loggerAdapter = loggerAdapter;
+        this.attributeConverters = attributeConverters;
+    }
+
+    protected RestLoggerAdapter restLoggerAdapter() {
+        RestLoggerAdapter adapter = configurer.map(PnetDataRestCallFactoryConfigurer::restLoggerAdapter).orElse(null);
+
+        if (adapter == null) {
+            adapter = loggerAdapter.orElse(null);
+        }
+
+        if (adapter == null) {
+            adapter = createDefaultRestLoggerAdapter();
+        }
+
+        return adapter;
+    }
+
+    private RestLoggerAdapter createDefaultRestLoggerAdapter() {
+        if (Slf4jRestLoggerAdapter.isSlf4jAvailable()) {
+            return Slf4jRestLoggerAdapter.getDefault();
+        }
+
+        return SystemRestLoggerAdapter.INSTANCE;
+    }
+
+    protected ObjectMapper objectMapper() {
+        return configurer
+            .map(PnetDataRestCallFactoryConfigurer::objectMapper)
+            .orElseGet(this::createDefaultObjectMapper);
+    }
+
+    private ObjectMapper createDefaultObjectMapper() {
         return JacksonPnetDataApiModule.createObjectMapper();
     }
 
-    @Bean
-    public ConversionService pnetDataApiConversionService(
-        Optional<Set<? extends Converter<?, ?>>> attributeConverters
-    ) {
+    protected ConversionService conversionService() {
+        return configurer
+            .map(PnetDataRestCallFactoryConfigurer::conversionService)
+            .orElseGet(this::createDefaultConversionService);
+    }
+
+    private ConversionService createDefaultConversionService() {
         ConversionServiceFactoryBean conversionServiceFactoryBean = new ConversionServiceFactoryBean();
 
-        attributeConverters.ifPresent(conversionServiceFactoryBean::setConverters);
+        Set<Converter<?, ?>> converters = new HashSet<>(attributeConverters.orElseGet(Set::of));
 
+        converters.add(localDateTimeToStringConverter());
+        converters.add(geoDistanceToStringConverter());
+
+        conversionServiceFactoryBean.setConverters(converters);
         conversionServiceFactoryBean.afterPropertiesSet();
 
         return conversionServiceFactoryBean.getObject();
     }
 
-    @Bean
-    public Converter<LocalDateTime, String> localDateTimeToStringConverter() {
+    private Converter<LocalDateTime, String> localDateTimeToStringConverter() {
         // do not convert this to a Lambda operation, otherwise Spring get's confused!
         return new Converter<>() {
             private final ZoneId systemDefault = ZoneId.systemDefault();
@@ -54,8 +102,7 @@ public class PnetDataRestCallFactoryConfig {
         };
     }
 
-    @Bean
-    public Converter<GeoDistance, String> geoDistanceToStringConverter() {
+    private Converter<GeoDistance, String> geoDistanceToStringConverter() {
         // do not convert this to a Lambda operation, otherwise Spring get's confused!
         return new Converter<>() {
             private final Locale locale = Locale.ENGLISH;