+++ pnet-data-api-java/src/main/java/pnet/data/api/resttemplate/RestTemplateBasedRestCallFactoryConfig.java
@@ -2,47 +2,45 @@ package pnet.data.api.resttemplate;
 
 import at.porscheinformatik.happyrest.RestCallFactory;
 import at.porscheinformatik.happyrest.RestLoggerAdapter;
-import at.porscheinformatik.happyrest.SystemRestLoggerAdapter;
-import at.porscheinformatik.happyrest.slf4j.Slf4jRestLoggerAdapter;
 import at.porscheinformatik.happyrest.spring.RestTemplateRestCallFactory;
 import at.porscheinformatik.happyrest.spring.SpringRestFormatter;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.Optional;
-import org.springframework.beans.factory.annotation.Qualifier;
+import java.util.Set;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
 import org.springframework.core.convert.ConversionService;
+import org.springframework.core.convert.converter.Converter;
 import org.springframework.http.client.SimpleClientHttpRequestFactory;
 import org.springframework.http.converter.HttpMessageConverter;
 import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
 import org.springframework.web.client.RestTemplate;
 import pnet.data.api.client.PnetDataRestCallFactoryConfig;
+import pnet.data.api.client.PnetDataRestCallFactoryConfigurer;
 import pnet.data.api.util.PnetDataApiUtils;
 
 @Configuration
-@Import({ PnetDataRestCallFactoryConfig.class })
-public class RestTemplateBasedRestCallFactoryConfig {
+public class RestTemplateBasedRestCallFactoryConfig extends PnetDataRestCallFactoryConfig {
+
+    protected RestTemplateBasedRestCallFactoryConfig(
+        Optional<PnetDataRestCallFactoryConfigurer> configurer,
+        Optional<RestLoggerAdapter> loggerAdapter,
+        Optional<Set<? extends Converter<?, ?>>> attributeConverters
+    ) {
+        super(configurer, loggerAdapter, attributeConverters);
+    }
 
     @Bean
-    public RestCallFactory restTemplateBasedRestCallFactory(
-        @Qualifier("pnetDataApiRestTemplate") RestTemplate restTemplate,
-        @Qualifier("pnetDataApiConversionService") ConversionService conversionService,
-        @Qualifier("pnetDataApiRestLoggerAdapter") Optional<RestLoggerAdapter> optionalLoggerAdapter
-    ) {
-        RestLoggerAdapter loggerAdapter = optionalLoggerAdapter.orElseGet(() -> {
-            if (Slf4jRestLoggerAdapter.isSlf4jAvailable()) {
-                return Slf4jRestLoggerAdapter.getDefault();
-            }
-
-            return SystemRestLoggerAdapter.INSTANCE;
-        });
+    public RestCallFactory restTemplateBasedRestCallFactory() {
+        RestTemplate restTemplate = restTemplate();
+        RestLoggerAdapter loggerAdapter = restLoggerAdapter();
+        ConversionService conversionService = conversionService();
 
         return new RestTemplateRestCallFactory(restTemplate, loggerAdapter, new SpringRestFormatter(conversionService));
     }
 
-    @Bean
-    public RestTemplate pnetDataApiRestTemplate(@Qualifier("pnetDataApiObjectMapper") ObjectMapper objectMapper) {
+    protected RestTemplate restTemplate() {
+        ObjectMapper objectMapper = objectMapper();
         RestTemplate restTemplate = new RestTemplate();
         SimpleClientHttpRequestFactory requestFactory =
             (SimpleClientHttpRequestFactory) restTemplate.getRequestFactory();